added basic read/write and CI
