+++ .gtihub/workflows/build_and_release.yaml
@@ -1,0 +1,47 @@
+name: Test and Release
+
+on:
+  push:
+    branches:
+      - main
+
+jobs:
+  Test-and-Release:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+      - uses: actions/setup-go@v3
+        with:
+          go-version: "1.19"
+      - name: Install GDAL Dev libs
+        run: |
+          sudo add-apt-repository ppa:ubuntugis/ppa
+          sudo apt-get update
+          sudo apt-get install -y libgdal-dev
+      - name: Run Tests
+        run: |
+          go test -short ./... -coverprofile=coverage.out -covermode=atomic
+      - name: Upload Coverage
+        uses: codecov/codecov-action@v3
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          fail_ci_if_error: true
+          files: ./coverage.out
+      - name: Increment Semver
+        id: semver
+        uses: paulhatch/semantic-version@v4.0.3
+        with:
+          tag_prefix: "v"
+          major_pattern: "(MAJOR)"
+          minor_pattern: "(MINOR)"
+          format: "${major}.${minor}.${patch}"
+          short_tags: false
+      - name: Create tag
+        id: tag
+        run: |
+          git config user.name github-actions
+          git config user.email gcp-headless@seerai.space
+          git tag ${{ steps.semver.outputs.version_tag }} -m "Beep, Boop. Written by a robot. Beep, Boop."
+          git push --tags
