+++ go.mod
@@ -1,3 +1,11 @@
 module github.com/seerai/godal
 
-go 1.15
+go 1.19
+
+require github.com/stretchr/testify v1.8.0
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
